% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_instances.R
\name{autofilter_instances}
\alias{autofilter_instances}
\title{autofilter_instances}
\usage{
autofilter_instances(
  data,
  site,
  year,
  geno,
  min_times = 2,
  na.rm = TRUE,
  max_cycles = 999
)
}
\arguments{
\item{data}{A *data.frame*.}

\item{site}{A *character* or *numeric* column in `data` representing site.}

\item{year}{A *character* or *numeric* column in `data` representing year.}

\item{geno}{A *character* or *numeric* column in `data` representing genotype.}

\item{min_times}{A *numeric* value for the minimum instances, e.g., minimum site-years.}

\item{na.rm}{A *logical* value. If TRUE, remove rows containing missing values.}

\item{max_cycles}{A *numeric* value for the maximum number of iterations for filtering.}
}
\value{
A *data.frame*, a filtered version of `data`.
}
\description{
Filters geno-years, site-years, and genos per site-year to
`min_times` instances.
}
