% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa_internal.R
\name{calculate_hfa}
\alias{calculate_hfa}
\title{calculate_hfa}
\usage{
calculate_hfa(x, ff, pheno, geno = NA, site = NA, year = NA)
}
\arguments{
\item{x}{data.frame}

\item{ff}{formula to use. part_pheno ~ geno */+ is_home}
}
\description{
calculate the home field advantage based on the formula, ff. Uses efficient
packages for fastest permutation results.
}
