% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_instances.R
\name{count_instances}
\alias{count_instances}
\title{count_instances}
\usage{
count_instances(data, group_col, rep_col, na.rm = FALSE)
}
\arguments{
\item{data}{A *data.frame*.}

\item{group_col}{A *character* or *numeric* column in `data` for primary grouping.}

\item{rep_col}{A *character* or *numeric* column in `data` indicating replicates of
`group_col`.}

\item{na.rm}{A *logical* value. If TRUE, remove rows containing missing values.}
}
\value{
A *vector* with the names of levels in `group_col` and the unique counts.
}
\description{
Count the number of times each category of `group_col` appears
per `rep_col`.
}
\seealso{
`filter_instances()`, `autofilter_instances()`
}
