% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_instances.R
\name{.autofilter_instances}
\alias{.autofilter_instances}
\title{Autofilter Instances}
\usage{
.autofilter_instances(
  data,
  site,
  year,
  geno,
  min_times = 2,
  na.rm = TRUE,
  max_cycles = 999
)
}
\arguments{
\item{data}{A *data.frame* or *data.table*.}

\item{site}{*character* or *numeric*: Column in `data` for site grouping.}

\item{year}{*character* or *numeric*: Column in `data` for year grouping.}

\item{geno}{*character* or *numeric*: Column in `data` for genotype grouping.}

\item{min_times}{*numeric*: Minimum number of instances (e.g., minimum site-years). Default is 2.}

\item{na.rm}{*logical*: Remove rows containing missing values? Default is `TRUE`.}

\item{max_cycles}{*numeric*: Maximum number of times to iteratively filter. Default is 999.}
}
\value{
A *data.frame* or *data.table*, a filtered version of `data`.
}
\description{
Filters geno-years, site-years, and genos per site-year to `min_times` instances.
}
\details{
Calls `.filter_instances()` iteratively for the combinations of {geno, year}, {site, year}, and {siteyear, geno}.
Repeats until no more rows are removed or `max_cycles` is reached.
`siteyear` is created using `paste(data[, site], data[, year], sep = "___")`.
}
\seealso{
`.filter_instances()`
}
