% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{.calculate_hfa}
\alias{.calculate_hfa}
\title{Calculate Home Field Advantage (HFA)}
\usage{
.calculate_hfa(x, ff, part_pheno, geno, year, LAPACK = TRUE)
}
\arguments{
\item{x}{A data frame containing the data to calculate HFA.}

\item{ff}{A formula to use (e.g., part_pheno ~ geno + is_home).}

\item{part_pheno}{A character string representing the phenotype column name.}

\item{geno}{A character string representing the genotype column name.}

\item{year}{A character string representing the year column name.}

\item{LAPACK}{A logical indicating whether to use LAPACK for QR decomposition (default is TRUE).}
}
\value{
A matrix of the home field advantage coefficients.
}
\description{
This function calculates the home field advantage based on the provided formula.
It uses efficient matrix operations for performance.
}
