% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{.calculate_intervals}
\alias{.calculate_intervals}
\title{Calculate Confidence Intervals for Permutation Results}
\usage{
.calculate_intervals(x)
}
\arguments{
\item{x}{A matrix of permutation results, where rows represent values and columns represent permutations.
The first column contains the observed data.}
}
\value{
A matrix with columns for the median, 5th percentile (q05), and 95th percentile (q95) of (observed - permutation).
}
\description{
This function calculates the median and 90% confidence intervals
(5th and 95th percentiles) for the difference between observed and permutation data.
}
