% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_hfa.R
\name{.calculate_temporal_hfa}
\alias{.calculate_temporal_hfa}
\title{Calculate Home Field Advantage (HFA) for Each Year}
\usage{
.calculate_temporal_hfa(x, ff, year)
}
\arguments{
\item{x}{A data frame containing the data.}

\item{ff}{A formula to use in the linear model, which should include the term <year>_num:is_home.}

\item{year}{A character string representing the year column in the formula.}
}
\value{
A data frame containing the HFA coefficients, standard errors, p-values, and adjusted p-values for each year.
}
\description{
This function calculates the home field advantage (HFA) coefficients for each year based on the specified formula.
}
\details{
The formula should contain a term such as <year>_num:is_home and specify a response variable (e.g., phenotype).
}
