% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_instances.R
\name{.count_instances}
\alias{.count_instances}
\title{Count Instances}
\usage{
.count_instances(data, group_col, rep_col, na.rm = FALSE)
}
\arguments{
\item{data}{A *data.frame* or *data.table*.}

\item{group_col}{*character* or *numeric*: Column in `data` for primary grouping.}

\item{rep_col}{*character* or *numeric*: Column in `data` indicating replicates of
`group_col`.}

\item{na.rm}{*logical* Remove rows containing missing values? Default is `FALSE`.}
}
\value{
A *named vector*. Names correspond to levels in `group_col` and values represent unique counts.
}
\description{
Count the number of times each category of `group_col` appears
per `rep_col`.
}
\seealso{
`.filter_instances()`, `.autofilter_instances()`
}
