% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_instances.R
\name{.filter_instances}
\alias{.filter_instances}
\title{Filter Instances}
\usage{
.filter_instances(data, group_col, rep_col, min_times = 2, na.rm = FALSE)
}
\arguments{
\item{data}{A *data.frame* or *data.table*.}

\item{group_col}{*character* or *numeric*: Column in `data` for primary grouping.}

\item{rep_col}{*character* or *numeric*: Column in `data` indicating replicates of `group_col`.}

\item{min_times}{*numeric*: Minimum number of instances (e.g., minimum site-years). Default is 2.}

\item{na.rm}{*logical*: Remove rows containing missing values? Default is `FALSE`.}
}
\value{
A *data.frame* or *data.table*, a filtered version of `data`.
}
\description{
Keep categories of `group_col` that occur in at least `min_times` of `rep_col` categories.
}
\seealso{
`.count_instances()`, `.autofilter_instances()`
}
