% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{.generate_sets}
\alias{.generate_sets}
\title{Generate Permutation Sets}
\usage{
.generate_sets(x, site, year, times, seed = NULL)
}
\arguments{
\item{x}{A data frame containing the data to permute.}

\item{site}{A character string identifying the column with spatial environment information.}

\item{year}{A character string identifying the column with temporal environment information.}

\item{times}{An integer indicating the number of permutation sets to generate.}

\item{seed}{An optional integer for setting the seed for reproducibility (default is NULL).}
}
\value{
A data frame of permutation sets, where row numbers are used to reorder `x`. The first column contains the original data order.
}
\description{
Generates structured permutation sets based on site-year combinations.
}
