% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_home.R
\name{.id_top_pheno}
\alias{.id_top_pheno}
\title{Identify Top Phenotype Site (Home Site)}
\usage{
.id_top_pheno(x, site, pheno, blup = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{A data.frame or matrix representing a single unit (e.g., a genotype). It is coerced to a data.frame.}

\item{site}{A character string specifying the column of the grouping variable (e.g., field site locations).}

\item{pheno}{A character string specifying the column of the response variable (e.g., yield, biomass).}

\item{blup}{Logical; if `TRUE`, uses shrinkage estimates from a linear mixed model.}

\item{verbose}{Logical; if `TRUE`, prints messages to the console about the process.}
}
\value{
The input data frame with an additional logical column `"is_home"`, indicating
whether each observation corresponds to the identified home site.
}
\description{
Identifies the level of `x[, site]` where the mean relative value of
`x[, pheno]` is highest. Uses shrinkage estimates of mean relative `x[, pheno]`
via `lme4` if `blup = TRUE` and if any site occurs in at least two years.
Otherwise, calculates simple means.
}
