% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_tests.R
\name{.one_tailed}
\alias{.one_tailed}
\title{.one_tailed <internal>}
\usage{
.one_tailed(x, direction = c("<=", ">"))
}
\arguments{
\item{x}{*matrix* where the first column is observations and the remaining
are permutations}

\item{direction}{*string* comparison formula, defaults to '<='}
}
\value{
a *vector* of the probability observations are different from
permutations in the `direction` indicated.
}
\description{
perform a one-tailed test comparing permuted to observed results
}
\details{
when `direction` = '<=', tests whether observations are *greater*
than expected. For HFA analysis, this would test for specialization.
}
\seealso{
`specialist_test()`, `generalist_test()`
}
