% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_home.R
\name{.scale_pheno}
\alias{.scale_pheno}
\title{Scale a Phenotype Column}
\usage{
.scale_pheno(x, pheno)
}
\arguments{
\item{x}{A data frame containing the data to scale.}

\item{pheno}{A string specifying the name of the column in `x` to scale (the phenotype).}
}
\value{
A data frame with an additional column containing the scaled values of the specified phenotype.
}
\description{
This function scales the values of a specified phenotype column in a data frame
by standardizing the data (z-scores), and then adds a new column with the scaled values.
The new column will have the name `rel_<pheno>`, where `<pheno>` is the name of the original phenotype.
}
