% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{.two_tailed}
\alias{.two_tailed}
\title{Perform Two-Tailed Test on Permutation Values}
\usage{
.two_tailed(x)
}
\arguments{
\item{x}{A matrix where each column represents a permutation, and the first column is the original data.}
}
\value{
A vector of p-values for each row in the matrix.
}
\description{
This function performs a two-tailed test on a matrix of permutation values.
Each column in the matrix represents the result of a different permutation, and the first column contains the original data.
}
