% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_instances.R
\name{filter_instances}
\alias{filter_instances}
\title{filter_instances}
\usage{
filter_instances(data, group_col, rep_col, min_times = 2, na.rm = FALSE)
}
\arguments{
\item{data}{A *data.frame*.}

\item{group_col}{A *character* or *numeric* column in `data` for primary grouping.}

\item{rep_col}{A *character* or *numeric* column in `data` indicating replicates of
`group_col`.}

\item{min_times}{A *numeric* value for the minimum number of instances, e.g., minimum site-years.}

\item{na.rm}{A *logical* value. If TRUE, remove rows containing missing values.}
}
\value{
A *data.frame*, a filtered version of `data`.
}
\description{
Keep categories of `group_col` that occur in at least
`min_times` of `rep_col` categories.
}
\seealso{
`count_instances()`, `autofilter_instances()`
}
