% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_tests.R
\name{generalist_test}
\alias{generalist_test}
\title{generalist_test}
\usage{
generalist_test(permute_results)
}
\arguments{
\item{permute_results}{*list*: the result from a call to `permute_hfa()`}
}
\value{
a *list*, modifying `permute_results` to include the test
in `permute_results$home_field['p_generalist']`. This is the *p*-value,
the probability of being called a generalist due to chance.
}
\description{
test whether a line is likely a generalist.
}
\details{
Runs a one-tailed test of whether observed home field advantage
is *less* than expected based on chance (permutations). Run `permute_hfa()`
first and feed the output into this function.
}
\seealso{
`permute_hfa()`, `specialist_test()`, `.one_tailed()`
}
