% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa_internal.R
\name{generate_sets}
\alias{generate_sets}
\title{generate_sets}
\usage{
generate_sets(x, site, year, times, seed = NULL)
}
\arguments{
\item{x}{data.frame}

\item{site}{column identifying spatial environment}

\item{year}{column identifying temporal environment}

\item{times}{number of sets to produce.}

\item{seed}{optional random seed for reproducibility.}
}
\value{
a data.frame of permutation sets - row numbers used to reorder x. The first column is the original data.
}
\description{
generate permutation sets structured by site-year
}
