% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home_site.R
\name{get_home_site}
\alias{get_home_site}
\title{get_home_site}
\usage{
get_home_site(data, geno, site)
}
\arguments{
\item{data}{data.frame from call to `id_home()`}

\item{geno}{column in `data` with genotype}

\item{site}{column in `data` with location environment}
}
\value{
a `data.frame` with columns `geno`, `site`, and `home_years` - the number
of years a genotype appeared at a home site.
}
\description{
extract home location information for each genotype
}
