% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libraries.R
\name{get_libraries}
\alias{get_libraries}
\title{try to load a library, and download if needed}
\usage{
get_libraries(lib_string, in_parallel = TRUE)
}
\arguments{
\item{lib_string}{charracter (vector) with package names}

\item{cores}{integer or logical - install packages in parallel? TRUE uses
`parallel::detectCores()`}
}
\value{
nothing, but loads packages into the user environment.
}
\description{
mostly helps resolve packrat bugs. Also installs in parallel.
}
