% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SS.R
\name{get_ss}
\alias{get_ss}
\title{get_ss}
\usage{
get_ss(model)
}
\arguments{
\item{model}{A data frame containing the model data.}
}
\value{
A data frame with the statistical measures
}
\description{
calculate the sum of squares for a model and present a summary of
the predictors, their sum of squares, percentage variance explained, F-values,
and p-values in a data frame. It relies on the Anova function from the car package.
}
