% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home_site.R
\name{home_distance_environmental}
\alias{home_distance_environmental}
\title{home_distance_environmental}
\usage{
home_distance_environmental(
  data,
  locations,
  geno,
  site,
  vars = c(),
  scale = TRUE,
  ...
)
}
\arguments{
\item{data}{A data.frame from a call to `id_home()`.}

\item{locations}{A data.frame with a column `site`.}

\item{geno}{Genotype column in `data`.}

\item{site}{Site environment column in `data` and `locations`.}

\item{vars}{A vector of columns in `locations` with environmental measurements.}

\item{scale}{A logical indicating whether `locations[, vars]` should be centered and scaled. Defaults to TRUE.}

\item{...}{Additional arguments passed to `dist()`.}
}
\value{
A distance matrix where rows and columns represent genotypes, and values represent distances
between the home sites of the genotype pairs.
}
\description{
Calculate the environmental distance between home sites of genotypes.
}
