% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home_site.R
\name{home_distance_spatial}
\alias{home_distance_spatial}
\title{home_distance_spatial}
\usage{
home_distance_spatial(
  data,
  locations,
  geno,
  site,
  lat = NA,
  long = NA,
  great_circle = TRUE
)
}
\arguments{
\item{data}{A data.frame from a call to `id_home()`.}

\item{locations}{A `SpatialPointsDataFrame` (recommended) or a data.frame with columns `site`, `lat`, and `long`.}

\item{geno}{Genotype column in `data` and in `locations`.}

\item{site}{Spatial environment column in `data`.}

\item{lat}{Latitude or y-coordinate column in `locations`.}

\item{long}{Longitude or x-coordinate column in `locations`.}

\item{great_circle}{A logical indicating whether coordinates are on an ellipse (set TRUE for lat/long).}
}
\value{
A distance matrix. Row and column names represent genotypes, and values represent distances
between the home sites of genotype pairs.
}
\description{
Calculate the distance between home sites for genotypes based on spatial locations.
}
\seealso{
`spDists` from the `"sp"` package.
}
