% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_home.R
\name{id_home}
\alias{id_home}
\title{Identify the Home Site Based on the Highest Relative Phenotype Value}
\usage{
id_home(
  df,
  site,
  year,
  geno,
  pheno,
  blup = TRUE,
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{df}{A data frame containing performance data by site, year, variety, etc.}

\item{site}{A character string indicating the column name for spatial units (e.g., field sites).}

\item{year}{A character string indicating the column name for temporal units (e.g., year).}

\item{geno}{A character string indicating the column name for genetic units (e.g., genotype).}

\item{pheno}{A character string indicating the column name for phenotype (e.g., yield, biomass).}

\item{blup}{A logical indicating whether to use shrinkage estimates (default is TRUE).}

\item{verbose}{A logical indicating whether to print messages (default is TRUE).}

\item{parallel}{A logical indicating whether to perform parallel processing (default is TRUE).}
}
\value{
A data frame with two new columns:
- `'rel_<pheno>'`: The relative phenotype value of each variety within each site-year.
- `'is_home'`: Logical column indicating whether a site is a variety's home site.
}
\description{
This function identifies the home site (location) where a given
variety performs best relative to other varieties across years. It can use
shrinkage estimates (BLUP) if required.
}
