% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_top_pheno.R
\name{id_top_pheno}
\alias{id_top_pheno}
\title{id_top_pheno}
\usage{
id_top_pheno(
  x,
  site,
  geno,
  pheno,
  method = c("blup", "mean", "median"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame or matrix for a single unit, ex a genotype. Coerced to data.frame.}

\item{site}{character of grouping column, ex field site locations}

\item{geno}{character, column name indicating genetic units}

\item{pheno}{character of response, ex yield, biomass}

\item{method}{character, method of finding highest relative performing location. Defaults to 'blup'. See details.}

\item{verbose}{logical (TRUE). Print messages?}

\item{...}{other arguments passed to method.}
}
\value{
x, with an additional logical column "is_home", which indicates
whether the value of x[, site] is the home site, and a column "method_rel_pheno", the
data used to identify home site.
}
\description{
id the level of x[, site] where the mean relative value of
x[, pheno] is highest. By default, uses shrinkage estimates of mean
relative x[, pheno] (via lme4) if any site occurs in at least two
years. Otherwise (or optionally) uses means.
}
\details{
method is passed to tapply(), so must return a single value per genotype-site combination.
}
