% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.rHFA}
\alias{print.rHFA}
\title{print.rHFA}
\usage{
\method{print}{rHFA}(x, ...)
}
\arguments{
\item{x}{object of `class("rHFA")`. Output from `permute_hfa()`.}
}
\description{
Print method for `class("rHFA")`
}
