% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_tests.R
\name{specialist_test}
\alias{specialist_test}
\title{specialist_test}
\usage{
specialist_test(permute_results)
}
\arguments{
\item{permute_results}{*list*: the result from a call to `permute_hfa()`}
}
\value{
a *list*, modifying `permute_results` to include the test
in `permute_results$home_field['p_specialist']`. This is the *p*-value,
the probability of being called a specialist due to chance.
}
\description{
test whether a line is likely a specialist.
}
\details{
Runs a one-tailed test of whether observed home field advantage
is *greater* than expected based on chance (permutations). Run `permute_hfa()`
first and feed the output into this function.
}
\seealso{
`permute_hfa()`, `generalist_test()`, `.one_tailed()`
}
