% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R, R/permute_hfa_internal.R
\name{.calculate_hfa}
\alias{.calculate_hfa}
\title{.calculate_hfa}
\usage{
.calculate_hfa(x, ff, pheno, geno = NA, site = NA, year = NA)

.calculate_hfa(x, ff, pheno, geno = NA, site = NA, year = NA)
}
\arguments{
\item{x}{data.frame}

\item{ff}{formula to use. part_pheno ~ geno */+ is_home}

\item{pheno}{character, column name indicating the phenotype (ex. yield)}

\item{geno}{character, column name indicating genetic units}

\item{site}{character, column name indicating spatial units}

\item{year}{character, column name indicating temporal units}
}
\value{
a data.frame of home_coefficients

a data.frame of home_coefficients
}
\description{
calculate the home field advantage based on the formula, ff. Uses efficient
packages for fastest permutation results.

calculate the home field advantage based on the formula, ff. Uses efficient
packages for fastest permutation results.
}
