% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{.calculate_intervals}
\alias{.calculate_intervals}
\title{.calculate_intervals <internal>}
\usage{
.calculate_intervals(x)
}
\arguments{
\item{x}{permutation results. Rows are values, columns are permutations. The first column is observed.}
}
\value{
a matrix of columns "median", "q05" (5th percentile), "q95" (95th percentile) of (observed - permutation)
}
\description{
calculate median and 90% confidence intervals of difference of permutation and observed data
}
