% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_hfa.R
\name{.calculate_temporal_hfa}
\alias{.calculate_temporal_hfa}
\title{calculate_temporal_hfa <internal>}
\usage{
.calculate_temporal_hfa(x, ff, year)
}
\arguments{
\item{x}{data.frame}

\item{ff}{formula}
}
\value{
a data.frame with the coefficients plus relevant
}
\description{
calculate hfa for each year
}
\details{
the formula should contain a term, <year>_num:is_home, and also specify a response (e.g. pheno)
}
