% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_home.R
\name{.id_top_pheno}
\alias{.id_top_pheno}
\title{.id_top_pheno}
\usage{
.id_top_pheno(x, site, pheno, blup = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{data.frame or matrix for a single unit, ex a genotype. Coerced to data.frame.}

\item{site}{character of grouping column, ex field site locations}

\item{pheno}{character of response, ex yield, biomass}

\item{blup}{logical - use shrinkage estimates?}

\item{verbose}{logical}
}
\value{
x, with an additional logical column `"Home"`, which indicates 
whether the value of `x[, site]` is the home site.
}
\description{
id the level of `x[, site]` where the mean relative value of 
`x[, pheno]` is highest. By default, uses shrinkage estimates of mean 
relative `x[, pheno]` (via `lme4`) if any site occurs in at least two 
years. Otherwise (or optionally) uses means.
}
