% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa_internal.R
\name{.quick_resid}
\alias{.quick_resid}
\title{.quick_resid, .quick_coef <internal>}
\usage{
.quick_resid(ff, data)
}
\arguments{
\item{ff}{formula}

\item{data}{data.frame containing formula}
}
\value{
A vector of residuals (coefficients).
}
\description{
calculate residuals (coefficients) of a linear model using Matrix, about an
order of magnitude faster than residuals(lm.fit()) for decently large, highly
sparse model matrices.
}
