% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{.quick_resid}
\alias{.quick_resid}
\title{.quick_resid <internal>}
\usage{
.quick_resid(ff, data)
}
\arguments{
\item{ff}{formula}

\item{data}{data.frame containing formula}
}
\value{
A vector of residuals, length = nrow(data)
}
\description{
calculate residuals of a linear model as efficiently as possible (using base R).
Residuals can be used in subsequent regressions as partial effects.
}
