% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{.two_tailed}
\alias{.two_tailed}
\title{.two_tailed <internal>}
\usage{
.two_tailed(x)
}
\arguments{
\item{x}{matrix. Each column is the result of a different permutation. The first column is the original data.}
}
\value{
a vector of p-values for each row in x.
}
\description{
perform two-tailed test on permutation values
}
