% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home_site.R
\name{get_home_site}
\alias{get_home_site}
\title{get_home_site}
\usage{
get_home_site(data, geno, site)
}
\arguments{
\item{data}{A data.frame from a call to `id_home()`.}

\item{geno}{A column in `data` with genotype.}

\item{site}{A column in `data` with location environment.}
}
\value{
A `data.frame` with columns `geno`, `site`, and `home_years`, which indicates the number
of years a genotype appeared at a home site.
}
\description{
Extract home location information for each genotype based on the results of `id_home()`.
}
