% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SS.R
\name{get_ss}
\alias{get_ss}
\title{get_ss}
\usage{
get_ss(model, test.statistic = "F", ...)
}
\arguments{
\item{model}{A model object or an ANOVA object.}

\item{test.statistic}{A character string indicating the test statistic to use. Defaults to 'F'.}

\item{...}{Additional arguments passed to `Anova()`.}
}
\value{
A data frame with the statistical measures, including sum of squares, percentage variance,
F-values, and p-values.
}
\description{
Calculate the sum of squares for a model and present a summary of
the predictors, their sum of squares, percentage variance explained, F-values,
and p-values in a data frame. It can work with both an ANOVA model or fit a model
using `car::Anova`.
}
