% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_home.R
\name{id_home}
\alias{id_home}
\title{id_home}
\usage{
id_home(
  data,
  site,
  year,
  geno,
  pheno,
  method = c("blup", "mean", "median"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{site}{character, column name indicating spatial units}

\item{year}{character, column name indicating temporal units}

\item{geno}{character, column name indicating genetic units}

\item{pheno}{character, column name indicating the phenotype (ex. yield)}

\item{method}{character, method of finding highest relative performing location. Defaults to 'blup'. See details.}

\item{verbose}{logical (TRUE). Print messages?}

\item{...}{other arguments passed to `method`.}

\item{df}{data.frame of performance data by site, year, and variety, for example.}
}
\value{
the data.frame `data` with three new columns:
- `'rel_<pheno>'`: numeric, the relative phenotype value of each variety within site-year
- `is_home`: logical of whether a site is a variety's home site.
- `<method>_rel_<pheno>`: numeric, the expected phenotype of each variety at that location
as derived by `method`.
}
\description{
Identify the home site based on highest relative phenotype value.
}
\details{
The home site for a given variety is defined as the location where
variety performs best across years, relative to other varieties. It's calculated
by:

1. Calculate relative phenotype within site-year by scaling (mean = 0) and
scaling (st. dev. = 1) phenotype across varieties within site-years.
2. Find the expected relative phenotype for each variety within
within sites across years.
3. For each variety, identify the site with the highest expected relative phenotype value.
}
\seealso{
`tapply()`, `lme4::lmer()`, `id_top_pheno()`
}
