% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_hfa.R
\name{permute_hfa}
\alias{permute_hfa}
\title{permute_hfa}
\usage{
permute_hfa(
  data,
  level = c("population", "genotype", "year", "site"),
  site = NA,
  year = NA,
  geno = NA,
  pheno = NA,
  popn = NA,
  covars = NA,
  times = 99,
  method = "blup",
  parallel = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{level}{parameter at which to apply the home field advantage. Options are
'population' (default), 'genotype', 'site', and 'year'.}

\item{site}{column name containing spatial environmental information}

\item{year}{column name containing temporal environmental information}

\item{geno}{column name containing genotype or variety information}

\item{pheno}{column name containing phenotype or performance information}

\item{popn}{column name differentiating sub-populations of genotypes within the data set.}

\item{covars}{formula or character. Specify covariates for the HFA analysis. See **Details**.}

\item{times}{how many times do you want to permute (in addition to observed data)?}

\item{method}{*character* for estimating relative performance at each location. Default is
'blup'. Other arguments that name functions (ex. `median`, `mean`, `quantile`) can be used.}

\item{parallel}{*logical* use parallel processing? Currently only on UNIX.}

\item{seed}{*numeric* for reproducible permutations}

\item{...}{added arguments passed to `method`.}
}
\value{
A list with four items: 

- `out$home_field` is a `data.frame` that summarizes the results. 
- `out$data` is a `data.frame` of the data used when calculating HFA. 
- `out$perms` is a matrix (or list of matrices) of permutation results
- `out$formula` is the formula used for calculating HFA
}
\description{
test the magnitude and significance of the home field advantage
versus what would be expected by chance.
}
\details{
This function is the only one that's needed for an HFA analysis. It identifies a home
site and calculates the home field advantage metric at the level chosen based on observed data.
It then permutes data within site-year and repeats this process to estimate the difference
between observed and expected (permuted) HFA - and tests whether this difference is zero.

Note that observed HFA may differ slightly from "manual" calculations (ex. via `lm()`)
as it uses sparse matrix operations for efficiency. See `Matrix::qr()` and `Matrix::qrR()`.

You can specify covariates to the HFA analysis - for example, environmental parameters if they are
included as columns in `data`. You must specify exactly how you want the covariates included - including
interactions with core HFA terms - using formula notation. For example, specifying `covars=~temperature + precipitation:is_home`
will allow preciptation, but not temperature, to interact with the home site designator. An equivalent
would be `covars="temperature + precipitation:is_home"`. You should probably check the formula output (`out$formula`).
}
\seealso{
`id_home()`
}
